// src/interfaces/details/starship-details.interface.ts
export interface StarhsipDetails {
  name: string;
  model: string;
  manufacturer: string;
  MGLT: number;
  length: number;
  max_atmosphering_speed: number;
  hyperdrive_rating: number;
  cargo_capacity: number;
  passengers: number;
  crew: string;
  consumables: string;
  cost_in_credits: number;
}